Minor refactoring - cleaning up the code.

+++ src/main/java/lt/lietpastas/partstore/repository/CountDTO.java
@@ -1,14 +1,16 @@
 package lt.lietpastas.partstore.repository;
 
+import java.math.BigDecimal;
+
 public class CountDTO {
 
-    private int value;
+    private BigDecimal value;
 
-    public int getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
 }
